+++ ompi/mca/common/ompio/Makefile.am
@@ -25,6 +25,7 @@ headers = \
 sources = \
 	common_ompio_print_queue.c \
 	common_ompio_file_open.c   \
+	common_ompio_file_view.c   \
 	common_ompio_file_read.c   \
 	common_ompio_file_write.c
