+++ ompi/mca/fcoll/base/Makefile.am
@@ -26,4 +26,5 @@ libmca_fcoll_la_SOURCES += \
         base/fcoll_base_file_select.c \
         base/fcoll_base_file_unselect.c \
         base/fcoll_base_find_available.c \
+        base/fcoll_base_sort.c \
         base/fcoll_base_coll_array.c
