+++ ompi/mca/io/ompio/io_ompio.c
@@ -38,93 +38,8 @@
 #include <math.h>
 #include <unistd.h>
 
-#ifdef HAVE_SYS_STATFS_H
-#include <sys/statfs.h> /* or <sys/vfs.h> */
-#endif
-#ifdef HAVE_SYS_PARAM_H
-#include <sys/param.h>
-#endif
-#ifdef HAVE_SYS_MOUNT_H
-#include <sys/mount.h>
-#endif
-#ifdef HAVE_SYS_STAT_H
-#include <sys/stat.h>
-#endif
 #include "io_ompio.h"
 
-int ompi_io_ompio_set_file_defaults (mca_io_ompio_file_t *fh)
-{
-
-   if (NULL != fh) {
-        ompi_datatype_t *types[2];
-        int blocklen[2] = {1, 1};
-        OPAL_PTRDIFF_TYPE d[2], base;
-        int i;
-
-        fh->f_io_array = NULL;
-        fh->f_perm = OMPIO_PERM_NULL;
-        fh->f_flags = 0;
-        fh->f_bytes_per_agg = mca_io_ompio_bytes_per_agg;
-        fh->f_datarep = strdup ("native");
-
-        fh->f_offset = 0;
-        fh->f_disp = 0;
-        fh->f_position_in_file_view = 0;
-        fh->f_index_in_file_view = 0;
-        fh->f_total_bytes = 0;
-
-        fh->f_init_procs_per_group = -1;
-        fh->f_init_procs_in_group = NULL;
-
-	fh->f_procs_per_group = -1;
-        fh->f_procs_in_group = NULL;
-
-        fh->f_init_num_aggrs = -1;
-        fh->f_init_aggr_list = NULL;
-
-
-        /* Default file View */
-        fh->f_iov_type = MPI_DATATYPE_NULL;
-        fh->f_stripe_size = mca_io_ompio_bytes_per_agg;
-	/*Decoded iovec of the file-view*/
-	fh->f_decoded_iov = NULL;
-        fh->f_etype = NULL;
-        fh->f_filetype = NULL;
-        fh->f_orig_filetype = NULL;
-
-	mca_io_ompio_set_view_internal(fh,
-				       0,
-				       &ompi_mpi_byte.dt,
-				       &ompi_mpi_byte.dt,
-                                       "native",
-				       fh->f_info);
-
-
-	/*Create a derived datatype for the created iovec */
-	types[0] = &ompi_mpi_long.dt;
-        types[1] = &ompi_mpi_long.dt;
-
-        d[0] = (OPAL_PTRDIFF_TYPE) fh->f_decoded_iov;
-        d[1] = (OPAL_PTRDIFF_TYPE) &fh->f_decoded_iov[0].iov_len;
-
-        base = d[0];
-        for (i=0 ; i<2 ; i++) {
-            d[i] -= base;
-        }
-
-        ompi_datatype_create_struct (2,
-                                     blocklen,
-                                     d,
-                                     types,
-                                     &fh->f_iov_type);
-        ompi_datatype_commit (&fh->f_iov_type);
-
-        return OMPI_SUCCESS;
-    }
-    else {
-        return OMPI_ERROR;
-    }
-}
 
 int ompi_io_ompio_generate_current_file_view (struct mca_io_ompio_file_t *fh,
                                               size_t max_data,
@@ -619,215 +534,7 @@ int ompi_io_ompio_decode_datatype (struct mca_io_ompio_file_t *fh,
     return OMPI_SUCCESS;
 }
 
-int ompi_io_ompio_sort (mca_io_ompio_io_array_t *io_array,
-                        int num_entries,
-                        int *sorted)
-{
-    int i = 0;
-    int j = 0;
-    int left = 0;
-    int right = 0;
-    int largest = 0;
-    int heap_size = num_entries - 1;
-    int temp = 0;
-    unsigned char done = 0;
-    int* temp_arr = NULL;
 
-    temp_arr = (int*)malloc(num_entries*sizeof(int));
-    if (NULL == temp_arr) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-    temp_arr[0] = 0;
-    for (i = 1; i < num_entries; ++i) {
-        temp_arr[i] = i;
-    }
-    /* num_entries can be a large no. so NO RECURSION */
-    for (i = num_entries/2-1 ; i>=0 ; i--) {
-        done = 0;
-        j = i;
-        largest = j;
-
-        while (!done) {
-            left = j*2+1;
-            right = j*2+2;
-            if ((left <= heap_size) &&
-                (io_array[temp_arr[left]].offset > io_array[temp_arr[j]].offset)) {
-                largest = left;
-            }
-            else {
-                largest = j;
-            }
-            if ((right <= heap_size) &&
-                (io_array[temp_arr[right]].offset >
-                 io_array[temp_arr[largest]].offset)) {
-                largest = right;
-            }
-            if (largest != j) {
-                temp = temp_arr[largest];
-                temp_arr[largest] = temp_arr[j];
-                temp_arr[j] = temp;
-                j = largest;
-            }
-            else {
-                done = 1;
-            }
-        }
-    }
-
-    for (i = num_entries-1; i >=1; --i) {
-        temp = temp_arr[0];
-        temp_arr[0] = temp_arr[i];
-        temp_arr[i] = temp;
-        heap_size--;
-        done = 0;
-        j = 0;
-        largest = j;
-
-        while (!done) {
-            left =  j*2+1;
-            right = j*2+2;
-
-            if ((left <= heap_size) &&
-                (io_array[temp_arr[left]].offset >
-                 io_array[temp_arr[j]].offset)) {
-                largest = left;
-            }
-            else {
-                largest = j;
-            }
-            if ((right <= heap_size) &&
-                (io_array[temp_arr[right]].offset >
-                 io_array[temp_arr[largest]].offset)) {
-                largest = right;
-            }
-            if (largest != j) {
-                temp = temp_arr[largest];
-                temp_arr[largest] = temp_arr[j];
-                temp_arr[j] = temp;
-                j = largest;
-            }
-            else {
-                done = 1;
-            }
-        }
-        sorted[i] = temp_arr[i];
-    }
-    sorted[0] = temp_arr[0];
-
-    if (NULL != temp_arr) {
-        free(temp_arr);
-        temp_arr = NULL;
-    }
-    return OMPI_SUCCESS;
-}
-
-int ompi_io_ompio_sort_iovec (struct iovec *iov,
-                              int num_entries,
-                              int *sorted)
-{
-    int i = 0;
-    int j = 0;
-    int left = 0;
-    int right = 0;
-    int largest = 0;
-    int heap_size = num_entries - 1;
-    int temp = 0;
-    unsigned char done = 0;
-    int* temp_arr = NULL;
-
-    if (0 == num_entries) {
-        return OMPI_SUCCESS;
-    }
-
-    temp_arr = (int*)malloc(num_entries*sizeof(int));
-    if (NULL == temp_arr) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-    temp_arr[0] = 0;
-    for (i = 1; i < num_entries; ++i) {
-        temp_arr[i] = i;
-    }
-    /* num_entries can be a large no. so NO RECURSION */
-    for (i = num_entries/2-1 ; i>=0 ; i--) {
-        done = 0;
-        j = i;
-        largest = j;
-
-        while (!done) {
-            left = j*2+1;
-            right = j*2+2;
-            if ((left <= heap_size) &&
-                (iov[temp_arr[left]].iov_base > iov[temp_arr[j]].iov_base)) {
-                largest = left;
-            }
-            else {
-                largest = j;
-            }
-            if ((right <= heap_size) &&
-                (iov[temp_arr[right]].iov_base >
-                 iov[temp_arr[largest]].iov_base)) {
-                largest = right;
-            }
-            if (largest != j) {
-                temp = temp_arr[largest];
-                temp_arr[largest] = temp_arr[j];
-                temp_arr[j] = temp;
-                j = largest;
-            }
-            else {
-                done = 1;
-            }
-        }
-    }
-
-    for (i = num_entries-1; i >=1; --i) {
-        temp = temp_arr[0];
-        temp_arr[0] = temp_arr[i];
-        temp_arr[i] = temp;
-        heap_size--;
-        done = 0;
-        j = 0;
-        largest = j;
-
-        while (!done) {
-            left =  j*2+1;
-            right = j*2+2;
-
-            if ((left <= heap_size) &&
-                (iov[temp_arr[left]].iov_base >
-                 iov[temp_arr[j]].iov_base)) {
-                largest = left;
-            }
-            else {
-                largest = j;
-            }
-            if ((right <= heap_size) &&
-                (iov[temp_arr[right]].iov_base >
-                 iov[temp_arr[largest]].iov_base)) {
-                largest = right;
-            }
-            if (largest != j) {
-                temp = temp_arr[largest];
-                temp_arr[largest] = temp_arr[j];
-                temp_arr[j] = temp;
-                j = largest;
-            }
-            else {
-                done = 1;
-            }
-        }
-        sorted[i] = temp_arr[i];
-    }
-    sorted[0] = temp_arr[0];
-
-    if (NULL != temp_arr) {
-        free(temp_arr);
-        temp_arr = NULL;
-    }
-    return OMPI_SUCCESS;
-}
 
 int ompi_io_ompio_sort_offlen (mca_io_ompio_offlen_array_t *io_array,
                                int num_entries,
