+++ ompi/mca/sharedfp/lockedfile/sharedfp_lockedfile_file_open.c
@@ -60,13 +60,13 @@ int mca_sharedfp_lockedfile_file_open (struct ompi_communicator_t *comm,
     data = (mca_io_ompio_data_t *) fh->f_fh->f_io_selected_data;
     ompio_fh = &data->ompio_fh;
 
-    err = mca_io_ompio_set_view_internal (shfileHandle,
-                                          ompio_fh->f_disp,
-                                          ompio_fh->f_etype,
-                                          ompio_fh->f_orig_filetype,
-                                          ompio_fh->f_datarep,
-                                          MPI_INFO_NULL);
-
+    err = mca_common_ompio_set_view (shfileHandle,
+                                     ompio_fh->f_disp,
+                                     ompio_fh->f_etype,
+                                     ompio_fh->f_orig_filetype,
+                                     ompio_fh->f_datarep,
+                                     MPI_INFO_NULL);
+    
 
     /*Memory is allocated here for the sh structure*/
     sh = (struct mca_sharedfp_base_data_t*)malloc(sizeof(struct mca_sharedfp_base_data_t));
